Test_data_sets template to choose the final one
Template of some standard procedures to decide which dataset we are going to choose. For each dataset you just need to change the name of the file in the second row of code and compute the next ones. -> it removes NaN values (including '?' values) and duplicates; shows the dataset shape before and after removing this rows; show some description and info about the dataset features and some histograms for each one